requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0

# LLM Provider APIs
openai>=1.0.0
anthropic>=0.18.0
google-generativeai>=0.3.0
requests>=2.31.0

# Metrics Analysis
vaderSentiment>=3.3.2
detoxify>=0.5.0

# Dashboard
streamlit>=1.28.0
altair>=5.0.0

